Remove background-color on focused input (usually submit buttons).
On firefox, this resize the button and may cause users to miss-click.
If this feature may have been previously added for accessibility (using
a color to stand out the focus), I believe the cost of miss-click is too
high to keep it. Moreover default style for accessibility improvement is
better left to the browser.